requirements.txt
# Core dependencies
torch>=1.9.1
torchvision>=0.15.0
opencv-python>=4.8.0
numpy==1.26.4
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0

# GPU acceleration
cupy-cuda12x>=12.0.0
faiss-gpu>=1.7.0
pyceres>=0.1.0

# Computer vision and feature extraction
kornia>=0.6.12
lightglue @ git+https://github.com/cvg/LightGlue.git
timm>=0.9.10

# JIT compilation and optimization
numba>=0.58.0

# Memory monitoring
psutil>=5.9.0

# Data storage and serialization
h5py>=3.8.0


# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
